+++ CHANGELOG.md
@@ -65,6 +65,7 @@
  * Added support for vertex effects. See `RaptorExample.cpp`.
  * Added ETC1 alpha support, thanks @halx99! Does not work when two color tint is enabled.
  * Added `spAtlasPage_setCustomTextureLoader()` which let's you do texture loading manually. Thanks @jareguo.
+ * Added `SkeletonRenderer:setSlotsRange()` and `SkeletonRenderer::createWithSkeleton()`. This allows you to split rendering of a skeleton up into multiple parts, and render other nodes in between. See `SkeletonRendererSeparatorExample.cpp` for an example.
 
 ### Cocos2d-Objc
  * Fixed renderer to work with 3.6 changes